Resolved conflicts in:
- core/src/joker_metadata.rs: Combined HEAD and main approaches
- core/src/joker_registry.rs: Used HEAD version with proper joker implementations
- core/src/shop/generation.rs: Used HEAD version for better test coverage
- pylatro/src/lib.rs: Used HEAD version with comprehensive joker metadata APIs
- pylatro/test/test_joker_metadata.py: Used HEAD version with comprehensive tests

+++ core/src/static_joker.rs
@@ -40,7 +40,7 @@ pub struct StaticJoker {
     /// Bonus mult to add
     pub mult_bonus: Option<i32>,
     /// Multiplier to apply to mult
-    pub mult_multiplier: Option<f32>,
+    pub mult_multiplier: Option<f64>,
     /// Condition for when to apply the effect
     pub condition: StaticCondition,
     /// Whether the effect applies per card or per hand
@@ -198,7 +198,7 @@ pub struct StaticJokerBuilder {
     base_cost: Option<usize>,
     chips_bonus: Option<i32>,
     mult_bonus: Option<i32>,
-    mult_multiplier: Option<f32>,
+    mult_multiplier: Option<f64>,
     condition: StaticCondition,
     per_card: bool,
 }
@@ -224,7 +224,7 @@ impl StaticJokerBuilder {
         self
     }
 
-    pub fn mult_multiplier(mut self, multiplier: f32) -> Self {
+    pub fn mult_multiplier(mut self, multiplier: f64) -> Self {
         self.mult_multiplier = Some(multiplier);
         self
     }
